+++ gpu-video/src/encoders.rs
@@ -1,4 +1,6 @@
-use crate::{EncodedOutputChunk, InputFrame, RawFrameData, VideoBackendError};
+use std::time::Duration;
+
+use crate::{InputFrame, RawFrameData, VideoBackendError};
 
 #[cfg(feature = "wgpu")]
 mod wgpu_api;
@@ -10,7 +12,10 @@ pub(crate) trait VideoEncoderBackend: Send {
         &mut self,
         frame: &InputFrame<RawFrameData>,
         force_idr: bool,
-    ) -> Result<EncodedOutputChunk<Vec<u8>>, VideoEncoderError>;
+        timeout: Duration,
+    ) -> Result<(), VideoEncoderError>;
+
+    fn flush(&mut self, timeout: Duration) -> Result<(), VideoEncoderError>;
 }
 
 pub(crate) trait VideoEncoderParametersInfoH264 {
@@ -42,17 +47,49 @@ pub struct BytesEncoderH264 {
 }
 
 impl BytesEncoderH264 {
-    /// The result is a chunk of H264 bitstream.
+    /// Encodes a frame. The resulting chunks of H264 bitstream are sent via the callback provided
+    /// at encoder creation.
     ///
     /// If the `force_keyframe` option is set to `true`, the encoder will encode this frame as a
     /// [keyframe](https://en.wikipedia.org/wiki/Video_compression_picture_types#Intra-coded_(I)_frames/slices_(key_frames)).
     /// Otherwise, the encoder will decide which frames should be coded this way.
+    ///
+    /// If [`EncoderOutputParameters::max_in_flight_submissions`](crate::parameters::EncoderOutputParameters::max_in_flight_submissions)
+    /// encode submissions are already in flight, this blocks until all submissions above the limit finish.
+    ///
+    /// Calling this from within the provided callback can lead to a deadlock.
     pub fn encode(
         &mut self,
         frame: &InputFrame<RawFrameData>,
         force_keyframe: bool,
-    ) -> Result<EncodedOutputChunk<Vec<u8>>, VideoEncoderError> {
-        self.encoder.encode_bytes(frame, force_keyframe)
+    ) -> Result<(), VideoEncoderError> {
+        self.encode_timeout(frame, force_keyframe, Duration::MAX)
+    }
+
+    /// Same as [`Self::encode`], but if [`EncoderOutputParameters::max_in_flight_submissions`](crate::parameters::EncoderOutputParameters::max_in_flight_submissions)
+    /// encode submissions are already in flight, this blocks until all submissions above the limit finish,
+    /// or times out after `timeout`.
+    pub fn encode_timeout(
+        &mut self,
+        frame: &InputFrame<RawFrameData>,
+        force_keyframe: bool,
+        timeout: Duration,
+    ) -> Result<(), VideoEncoderError> {
+        self.encoder.encode_bytes(frame, force_keyframe, timeout)
+    }
+
+    /// Flush all chunks from the encoder.
+    /// This blocks until all chunks have been sent via the provided callback.
+    ///
+    /// Calling this from within the provided callback can lead to a deadlock.
+    pub fn flush(&mut self) -> Result<(), VideoEncoderError> {
+        self.flush_timeout(Duration::MAX)
+    }
+
+    /// Flush all chunks from the encoder.
+    /// This blocks until all chunks have been sent via the provided callback, or times out after `timeout`.
+    pub fn flush_timeout(&mut self, timeout: Duration) -> Result<(), VideoEncoderError> {
+        self.encoder.flush(timeout)
     }
 
     /// Retrieve encoded SPS NAL units from the video session parameters, in Annex B.
@@ -78,17 +115,49 @@ pub struct BytesEncoderH265 {
 }
 
 impl BytesEncoderH265 {
-    /// The result is a chunk of H265 bitstream.
+    /// Encodes a frame. The resulting chunks of H265 bitstream are sent via the callback provided
+    /// at encoder creation.
     ///
     /// If the `force_keyframe` option is set to `true`, the encoder will encode this frame as a
     /// [keyframe](https://en.wikipedia.org/wiki/Video_compression_picture_types#Intra-coded_(I)_frames/slices_(key_frames)).
     /// Otherwise, the encoder will decide which frames should be coded this way.
+    ///
+    /// If [`EncoderOutputParameters::max_in_flight_submissions`](crate::parameters::EncoderOutputParameters::max_in_flight_submissions)
+    /// encode submissions are already in flight, this blocks until all submissions above the limit finish.
+    ///
+    /// Calling this from within the provided callback can lead to a deadlock.
     pub fn encode(
         &mut self,
         frame: &InputFrame<RawFrameData>,
         force_keyframe: bool,
-    ) -> Result<EncodedOutputChunk<Vec<u8>>, VideoEncoderError> {
-        self.encoder.encode_bytes(frame, force_keyframe)
+    ) -> Result<(), VideoEncoderError> {
+        self.encode_timeout(frame, force_keyframe, Duration::MAX)
+    }
+
+    /// Same as [`Self::encode`], but if [`EncoderOutputParameters::max_in_flight_submissions`](crate::parameters::EncoderOutputParameters::max_in_flight_submissions)
+    /// encode submissions are already in flight, this blocks until all submissions above the limit finish,
+    /// or times out after `timeout`.
+    pub fn encode_timeout(
+        &mut self,
+        frame: &InputFrame<RawFrameData>,
+        force_keyframe: bool,
+        timeout: Duration,
+    ) -> Result<(), VideoEncoderError> {
+        self.encoder.encode_bytes(frame, force_keyframe, timeout)
+    }
+
+    /// Flush all chunks from the encoder.
+    /// This blocks until all chunks have been sent via the provided callback.
+    ///
+    /// Calling this from within the provided callback can lead to a deadlock.
+    pub fn flush(&mut self) -> Result<(), VideoEncoderError> {
+        self.flush_timeout(Duration::MAX)
+    }
+
+    /// Flush all chunks from the encoder.
+    /// This blocks until all chunks have been sent via the provided callback, or times out after `timeout`.
+    pub fn flush_timeout(&mut self, timeout: Duration) -> Result<(), VideoEncoderError> {
+        self.encoder.flush(timeout)
     }
 
     /// Retrieve encoded VPS NAL units from the video session parameters, in Annex B.
@@ -147,6 +216,9 @@ pub enum VideoEncoderError {
     #[cfg(feature = "wgpu")]
     #[error(transparent)]
     WgpuTextureEncoderError(#[from] WgpuTextureEncoderError),
+
+    #[error("Encode submission timed out")]
+    EncodeSubmissionTimeout,
 
     #[error("Encoder error: {0}")]
     BackendError(VideoBackendError),