+++ gpu-video/CHANGELOG.md
@@ -20,6 +20,7 @@
 ### ✨ New features
 - Added VUI support to H.265 encoder ([#1995](https://github.com/software-mansion/smelter/pull/1995) by @noituri)
 - Made decoders work asynchronously. Bytes decoder now returns frames via callback ([#2143](https://github.com/software-mansion/smelter/pull/2143) by @noituri)
+- Made encoders work asynchronously. Encoded frames are returned via callback ([#2215](https://github.com/software-mansion/smelter/pull/2215) by @noituri)
 
 ### 🐛 Bug fixes
 - Request only the transfer and compute queues that are actually used when creating the device. ([#2223](https://github.com/software-mansion/smelter/pull/2223) by @jerzywilczek)